better file organisation and added HISTORY.md
