+++ HISTORY.md
@@ -1,0 +1,3 @@
+### Aibox 0.1.0
+
+* Adding pretrained models from `transformers` and add relevant tests in a new branch called `nlp`
