+++ aibox/load.py
@@ -1,4 +1,3 @@
-
 import transformers
 from efficientnet_pytorch import EfficientNet
