+++ tests/assets/cat.jpg
