+++ try.py
@@ -1,0 +1,4 @@
+import aibox
+
+model = aibox.load("mobilenetv2")
+out = model('./tests/assets/cat.jpg')
